Fix win64 build.
git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@8665 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/ast.h
@@ -138,7 +138,7 @@ class AstNode: public ZoneObject {
   void* operator new(size_t size, Zone* zone) {
     Isolate* isolate = zone->isolate();
     isolate->set_ast_node_count(isolate->ast_node_count() + 1);
-    return zone->New(size);
+    return zone->New(static_cast<int>(size));
   }
 
   AstNode() {}
